requirements.txt
# Django Core
Django==4.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Database & Cache
django-redis==5.4.0
redis==5.0.1

# REST API
djangorestframework==3.14.0
django-filter==23.5
drf-yasg==1.21.7

# Authentication & Security
django-cors-headers==4.3.1

# Document Processing
Pillow==10.1.0
reportlab==4.0.7
python-docx==1.1.0
openpyxl==3.1.2

# Tasks & Scheduling
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Notifications (исправленная версия)
#django-telegram-bot==0.6.0
requests==2.31.0

# Import/Export
django-import-export==3.3.6

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Monitoring & Debugging
django-debug-toolbar==4.2.0
django-silk==5.0.3
django-health-check==3.17.0

# Logging
loguru==0.7.2

# Encryption
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3